requirements.txt
python==3.7.7
pytorch==1.6.0
torchvision==0.7.0
torchtext==0.7
numpy==1.19
dill==0.3.3
msgpack-numpy==0.4.7.1
msgpack-python==1.0.2
spacy==2.3.5
tensorboard==1.14.0
tensorflow==1.14.0
terminado==0.9.2
tqdm==4.56.0